+++ backend/genesis-core/src/ym2612.rs
@@ -23,6 +23,7 @@ const FM_SAMPLE_DIVIDER: u8 = 24;
 
 // Phase is 10 bits
 const PHASE_MASK: u16 = 0x03FF;
+const HALF_PHASE_MASK: u16 = PHASE_MASK >> 1;
 
 // Operator output is signed 14-bit
 const OPERATOR_OUTPUT_MIN: i16 = -0x2000;
@@ -81,7 +82,6 @@ impl FmOperator {
         }
     }
 
-    // TODO optimize to avoid floating-point arithmetic
     fn sample_clock(&mut self, modulation_input: u16) -> i16 {
         let feedback = match self.feedback_level {
             0 => 0,
@@ -95,22 +95,28 @@ impl FmOperator {
         };
 
         let phase = (self.phase.current_phase() + modulation_input + feedback) & PHASE_MASK;
-        let sine = phase_sin(phase);
+
+        // Phase is a 10-bit value that represents a number in the range 0 to 2*PI.
+        // Actual hardware splits this into a sign bit and a half-phase value from 0 to PI, computes
+        // the amplitude based on the half-phase, and then applies the sign bit at final output
+        let sign = phase.bit(9);
+        let sine_attenuation = phase_to_attenuation(phase);
 
         let envelope_attenuation = self.envelope.current_attenuation();
-        let attenuation = if self.am_enabled {
+        let envelope_am_attenuation = if self.am_enabled {
             let am_attenuation = lfo::amplitude_modulation(self.lfo_counter, self.am_sensitivity);
             (envelope_attenuation + am_attenuation).clamp(0, envelope::MAX_ATTENUATION)
         } else {
             envelope_attenuation
         };
 
-        // Convert from attenuation in dB to volume in linear units
-        let volume = attenuation_to_volume(attenuation);
-        let amplitude = sine * volume;
+        // Add phase attenuation (4.8 fixed-point) and envelope/AM attenuation (4.6 fixed-point)
+        let total_attenuation = sine_attenuation + (envelope_am_attenuation << 2);
 
-        // Convert volume to a 14-bit signed integer representing a floating-point value between -1 and 1
-        let output = (amplitude * f64::from(OPERATOR_OUTPUT_MAX)).round() as i16;
+        // Compute final output, adding the sign bit back in
+        let amplitude = attenuation_to_amplitude(total_attenuation);
+        let output = if sign { -(amplitude as i16) } else { amplitude as i16 };
+
         self.last_output = self.current_output;
         self.current_output = output;
 
@@ -118,27 +124,62 @@ impl FmOperator {
     }
 }
 
+// Logic based on http://gendev.spritesmind.net/forum/viewtopic.php?p=6114#p6114
 #[inline]
-fn phase_sin(phase: u16) -> f64 {
-    // Phase represents a value from 0 to 2PI on a scale from 0 to 2^10
-    static LOOKUP_TABLE: LazyLock<[f64; 1024]> = LazyLock::new(|| {
-        array::from_fn(|i| (i as f64 / 1024.0 * 2.0 * std::f64::consts::PI).sin())
-    });
+fn phase_to_attenuation(phase: u16) -> u16 {
+    // Actual hardware has a 256-entry quarter-sine table. This is emulated using a half-sine table
+    // for simplicity, but the values are calculated the same way
+    static LOG_SINE_TABLE: LazyLock<[u16; 512]> = LazyLock::new(|| {
+        array::from_fn(|mut i| {
+            use std::f64::consts::PI;
 
-    LOOKUP_TABLE[phase as usize]
-}
+            if i.bit(8) {
+                // Second quarter-phase
+                i = (!i) & 0xFF;
+            }
 
-#[inline]
-fn attenuation_to_volume(attenuation: u16) -> f64 {
-    // Envelope attenuation represents a value from 0dB to 96dB on a scale from 0 to 2^10
-    static LOOKUP_TABLE: LazyLock<[f64; 1024]> = LazyLock::new(|| {
-        array::from_fn(|i| {
-            let decibels = 96.0 * i as f64 / 1024.0;
-            10.0_f64.powf(decibels / -20.0)
+            // The table indices represent numbers in the range 0 to PI/2, but slightly offset in order
+            // to avoid computing log2(0)
+            let n = ((i << 1) | 1) as f64;
+            let sine = (n / 512.0 * PI / 2.0).sin();
+
+            // The table stores attenuation values, but on a log2 scale instead of log10
+            let attenuation = -sine.log2();
+
+            // Table contains 12-bit values that represent 4.8 fixed-point
+            (attenuation * f64::from(1 << 8)).round() as u16
         })
     });
 
-    LOOKUP_TABLE[attenuation as usize]
+    LOG_SINE_TABLE[(phase & HALF_PHASE_MASK) as usize]
+}
+
+// Logic based on http://gendev.spritesmind.net/forum/viewtopic.php?p=6114#p6114
+#[inline]
+fn attenuation_to_amplitude(attenuation: u16) -> u16 {
+    static POW2_TABLE: LazyLock<[u16; 256]> = LazyLock::new(|| {
+        array::from_fn(|i| {
+            // This is a lookup table for 2^(-n), where n is a value between 0 and 1
+            // Index i represents the number (i + 1)/256
+            let n = ((i + 1) as f64) / 256.0;
+            let inverse_pow2 = 2.0_f64.powf(-n);
+
+            // Table contains 11-bit values that represent 0.11 fixed-point
+            (inverse_pow2 * f64::from(1 << 11)).round() as u16
+        })
+    });
+
+    // Attenuation is interpreted as a 5.8 fixed-point number on a log2 scale
+    let int_part = (attenuation >> 8) & 0x1F;
+    if int_part >= 13 {
+        // Final result is guaranteed to shift down to 0
+        // Int part is applied as a right shift to 13-bit values
+        return 0;
+    }
+
+    let fract_part = attenuation & 0xFF;
+    let fract_pow2 = POW2_TABLE[fract_part as usize];
+    (fract_pow2 << 2) >> int_part
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Default, Encode, Decode)]